Source/Phase3/requirements.txt
fastapi 
torch 
torchvision 
uvicorn 
pillow 
python-multipart 
requests
psutil
mlflow
starlette
keras
tensorflow
numpy
tensorflow-addons
scikit-image